Update vector alert content #2884
